cli: long-syntax support for `--volume` (deprecates `--mount` for `docker service`)
The long-syntax is compatible to `--mount` except that the flag name is `--volume`.

 * `docker run`:     supports both the short-syntax and the long-syntax
 * `docker service`: only supports the long-syntax

The long/short detection is done by using some regexp matching implemented in
`opts/volume.go`. (see `InferVolumeType()`)

This commit also introduces `--volume-add` and `--volume-rm` for `docker service`,
and hence deprecates `--mount-add` and `--mount-rm`.
Note that `--mount`, `--mount-add`, and `--mount-rm` are still available for
compatibility.

Signed-off-by: Akihiro Suda <suda.akihiro@lab.ntt.co.jp>