move rootless-check to initFirewalld(), log as "info"
As this is an expected scenario, it's better to skip the detection and log, than
to return an error (and ignore / log as debug info).

Perhaps it would be cleaner if we pass the "rootless" mode as an argument to
`GetIptable()` instead (initialise with these options), but there's a lot more
to clean up here.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>